pyproject.toml
[project]
name = "haintech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marcin Leliwa", email = "marcin.leliwa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ampf>=0.1.18",
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "openai>=1.68.2",
    "pydantic>=2.10.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.11.2",
]

[[tool.uv.index]]
name = "private-registry"
url = "https://europe-west3-python.pkg.dev/development-428212/pip"
publish-url = "https://europe-west3-python.pkg.dev/development-428212/pip"